# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "All HTTP + MQ endpoints exposed by the goop2 viewer.\\n\\nAll peer-to-peer signaling travels through MQ (POST /api/mq/send → P2P → /api/mq/events SSE).\\nSee x-mq-topics in the raw spec (/api/openapi.json) for the full topic contract.",
        "title": "goop2 Viewer API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/templates/apply": {
            "post": {
                "description": "Resets the site and database, then applies the named built-in template.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Apply a built-in template",
                "parameters": [
                    {
                        "description": "Template to apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyResponse"
                        }
                    },
                    "400": {
                        "description": "template name required / template not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "bad csrf",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/templates/apply-local": {
            "post": {
                "description": "Reads all files from the given directory (must contain manifest.json), resets site and database, then applies the template.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Apply a template from a local folder",
                "parameters": [
                    {
                        "description": "Folder path and CSRF",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyLocalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyResponse"
                        }
                    },
                    "400": {
                        "description": "path required / invalid path / not a directory / manifest.json not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "bad csrf",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/templates/apply-store": {
            "post": {
                "description": "Spends credits if required, downloads the template bundle from the rendezvous server, resets site and database, then applies the template.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Apply a store template (download, spend credits, apply)",
                "parameters": [
                    {
                        "description": "Store template to apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.templateApplyStoreResponse"
                        }
                    },
                    "400": {
                        "description": "template name required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "insufficient credits",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "bad csrf",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "failed to download template",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/templates/validate-local": {
            "post": {
                "description": "Reads manifest.json from the given absolute path and returns its metadata for preview.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Validate a local template folder",
                "parameters": [
                    {
                        "description": "Folder path",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.templateValidateLocalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.templateValidateLocalResponse"
                        }
                    },
                    "400": {
                        "description": "path required / not a directory / manifest.json not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/templates/{dir}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Fetch template bundle or manifest by directory name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template directory name",
                        "name": "dir",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.templateApplyLocalRequest": {
            "type": "object",
            "properties": {
                "csrf": {
                    "type": "string",
                    "example": "token123"
                },
                "path": {
                    "type": "string",
                    "example": "/home/user/my-template"
                }
            }
        },
        "routes.templateApplyRequest": {
            "type": "object",
            "properties": {
                "csrf": {
                    "type": "string",
                    "example": "token123"
                },
                "template": {
                    "type": "string",
                    "example": "corkboard"
                }
            }
        },
        "routes.templateApplyResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "applied"
                },
                "template": {
                    "type": "string",
                    "example": "corkboard"
                }
            }
        },
        "routes.templateApplyStoreRequest": {
            "type": "object",
            "properties": {
                "csrf": {
                    "type": "string",
                    "example": "token123"
                },
                "template": {
                    "type": "string",
                    "example": "kanban"
                }
            }
        },
        "routes.templateApplyStoreResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 450
                },
                "status": {
                    "type": "string",
                    "example": "applied"
                },
                "template": {
                    "type": "string",
                    "example": "kanban"
                }
            }
        },
        "routes.templateValidateLocalRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string",
                    "example": "/home/user/my-template"
                }
            }
        },
        "routes.templateValidateLocalResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "blog"
                },
                "description": {
                    "type": "string",
                    "example": "A custom template"
                },
                "icon": {
                    "type": "string",
                    "example": "pencil"
                },
                "name": {
                    "type": "string",
                    "example": "My Template"
                }
            }
        },
...
    }
}